Make use of Google Maps Places API
Four modes to try to control and retain API credits usage.
Change a value of the `GAPI_MODE` environment variable to switch between them.

```rust
pub enum GoogleAPIMode {
    Place,      // Find Place request
    Text,       // Text Search request
    GeoPlace,   // Geocoding request first, Find Place if ZERO_RESULTS
    GeoText,    // Geocoding request first, Text Search if ZERO_RESULTS
}
```

By default, the `GeoPlace` mode is used. If it's results is insufficient, try `GeoText`.
The *Text Search* responses with a lot more rich result list.
However, they both execute 2 requests in the worst case. If this is the most often case, switch to the `Place` or `Text` mode.